FadeAndDestroy naar base. FamiliarStates in aparte class
