+++ Assets/Scripts/AI/BaseEnemyAI.cs
@@ -27,8 +27,6 @@ public abstract class BaseEnemyAI : BaseUnitAI
 
     protected float _distanceToTarget;
 
-    protected float _fadeTime = 0.2f;
-
     protected override void Start()
     {
         UnitManager.EnemyUnits.Add(transform);
@@ -149,25 +147,5 @@ public abstract class BaseEnemyAI : BaseUnitAI
 
         // fade sprite out
         StartCoroutine(FadeOutAndDestroy());
-    }
-
-    private IEnumerator FadeOutAndDestroy ()
-    {
-        Color initialColor = Sprite.color;
-        float t = 0f;
-        while (t < _fadeTime * 0.5f)
-        {
-            t += Time.deltaTime;
-            Sprite.color = Color.Lerp(initialColor, Color.black, t * 2f / _fadeTime);
-            yield return null;
-        }
-        t = 0f;
-        while (t < _fadeTime * 0.5f)
-        {
-            t += Time.deltaTime;
-            Sprite.color = Color.Lerp(Color.black, Color.clear, t * 2f / _fadeTime);
-            yield return null;
-        }
-        Destroy(gameObject);
     }
 }
