+++ Assets/Scripts/AI/States/FamiliarCombatStates.cs
@@ -1,0 +1,13 @@
+public enum FamiliarCombatState 
+{
+    Following,
+    Agro,
+    Dead
+}
+
+public enum FamiliarAgroStates
+{
+    Chase,
+
+    Attack
+}
