+++ Assets/Scripts/AI/States/FamiliarCombatStates.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9914aa93d32b73b479148cd8927e8718
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
